Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-iframe-element/sandbox_032-expected.txt

CONSOLE MESSAGE: Origin null is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://localhost:8800/html/semantics/embedded-content/the-iframe-element/support/standalone-pass.htm due to access control checks.

PASS Block XMLHttpRequest in an iframe with the sandbox attribute.
